+++ pganonymizer/providers.py
@@ -1,3 +1,4 @@
+import operator
 import random
 from hashlib import md5
 
@@ -94,9 +95,9 @@ class FakeProvider(with_metaclass(ProviderMeta, Provider)):
         return cls.id.lower() == name.split('.')[0].lower()
 
     def alter_value(self, value):
-        func_name = self.kwargs['name'].split('.')[1]
+        func_name = self.kwargs['name'].split('.', 1)[1]
         try:
-            func = getattr(fake_data, func_name)
+            func = operator.attrgetter(func_name)(fake_data)
         except AttributeError as exc:
             raise InvalidProviderArgument(exc)
         return func()