+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
-from mock import Mock, call, patch
+from mock import ANY, Mock, call, patch
 
-from pganonymizer.utils import get_connection, truncate_tables
+from pganonymizer.utils import get_connection, import_data, truncate_tables
 
 
 class TestGetConnection:
@@ -38,3 +38,24 @@ class TestTruncateTables:
             connection.cursor.assert_not_called()
             mock_cursor.execute.assert_not_called()
             mock_cursor.close.assert_not_called()
+
+
+class TestImportData:
+
+    @pytest.mark.parametrize('source_table, table_columns, primary_key, expected_tbl_name, expected_columns', [
+        ['src_tbl', ['id', 'COL_1'], 'id', 'tmp_src_tbl', ['"id"', '"COL_1"']]
+    ])
+    def test(self, source_table, table_columns, primary_key, expected_tbl_name, expected_columns):
+        mock_cursor = Mock()
+
+        connection = Mock()
+        connection.cursor.return_value = mock_cursor
+
+        import_data(connection, {}, source_table, table_columns, primary_key, [])
+
+        assert connection.cursor.call_count == 2
+        assert mock_cursor.close.call_count == 2
+
+        mock_cursor.copy_from.assert_called_once()
+        expected = [call(ANY, expected_tbl_name, columns=expected_columns, null=ANY, sep=ANY)]
+        assert mock_cursor.copy_from.call_args_list == expected