+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Simplified the CLI database connection arguments
 
+## 0.5.0 (2021-06-30)
+
+* [#22](https://github.com/rheinwerk-verlag/postgresql-anonymizer/pull/22): Fix table and column name quotes in cursor.copy_from call ([nurikk](https://github.com/nurikk))
+* [#23](https://github.com/rheinwerk-verlag/postgresql-anonymizer/pull/23): Allow uniq faker ([nurikk](https://github.com/nurikk))
+
 ## 0.4.1 (2021-05-27)
 
 * [#19](https://github.com/rheinwerk-verlag/postgresql-anonymizer/pull/19): Make chunk size in the table definition dynamic ([halilkaya](https://github.com/halilkaya))